+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Input box for Value +- is usable even when strictly integer types are selected
 - Integer search with Value +- near integer limits works correctly instead of rolling over and giving no results
 - Live value update thread sleeps proportionately to how long it takes to update values, should use ~5% CPU time on that core
+- Decide to search with certain value type including deviation edges